+++ application/datamapper/htmlform.php
@@ -31,8 +31,8 @@ if (!function_exists('dmz_log_message')) {
  * @package		DMZ-Included-Extensions
  * @category	DMZ
  * @author  	Phil DeJarnett
- * @link    	http://www.overzealous.com/dmz/pages/extensions/htmlform.html
- * @version 	1.0
+ * @link    	https://github.com/P2GR/datamapper2
+ * @version 	2.0.0
  */
 
 // --------------------------------------------------------------------------