+++ application/datamapper/simplecache.php
@@ -9,8 +9,8 @@
  * @package		DMZ-Included-Extensions
  * @category	DMZ
  * @author  	Phil DeJarnett
- * @link    	http://www.overzealous.com/dmz/pages/extensions/simplecache.html
- * @version 	1.0
+ * @link    	https://github.com/P2GR/datamapper2
+ * @version 	2.0.0
  */
 
 // --------------------------------------------------------------------------