+++ application/datamapper/translate.php
@@ -9,7 +9,8 @@
  * @package		DMZ-Included-Extensions
  * @category	DMZ
  * @author  	WanWizard
- * @version 	1.0
+ * @link    	https://github.com/P2GR/datamapper2
+ * @version 	2.0.0
  */
 
 // --------------------------------------------------------------------------